feat(tests): separate app and server in different files
