+++ package.json
@@ -6,7 +6,7 @@
     "lint": "eslint .",
     "start": "node bin/server.js",
     "dev": "nodemon bin/server.js",
-    "test": "NODE_ENV=test jest --forceExit --verbose"
+    "test": "jest --forceExit --verbose"
   },
   "dependencies": {
     "@rama41222/node-logger": "^1.0.1",
